+++ datacube/index/abstract.py
@@ -16,6 +16,7 @@ from typing import (Any, Iterable, Iterator,
                     NamedTuple, Optional,
                     Tuple, Union, Sequence)
 from uuid import UUID
+from datetime import timedelta
 
 from datacube.config import LocalConfig
 from datacube.index.exceptions import TransactionException
@@ -904,10 +905,27 @@ class AbstractDatasetResource(ABC):
 
         :param Dataset ds: dataset to search
         """
+        less_mature = self.find_less_mature(ds)
+        less_mature_ids = map(lambda x: x.id, less_mature)
+
+        self.archive(less_mature_ids)
+        for lm_ds in less_mature_ids:
+            _LOG.info(f"Archived less mature dataset: {lm_ds}")
+
+    def find_less_mature(self, ds: Dataset) -> Iterable[Dataset]:
+        """
+        Find less mature versions of a dataset
+
+        :param Dataset ds: Dataset to search
+        :return: Iterable of less mature datasets
+        """
         less_mature = []
+        # 'expand' the date range by a millisecond to give a bit more leniency in datetime comparison
+        expanded_time_range = Range(ds.metadata.time.begin - timedelta(milliseconds=1),
+                                    ds.metadata.time.end + timedelta(milliseconds=1))
         dupes = self.search(product=ds.product.name,
                             region_code=ds.metadata.region_code,
-                            time=ds.metadata.time)
+                            time=expanded_time_range)
         for dupe in dupes:
             if dupe.id == ds.id:
                 continue
@@ -924,10 +942,8 @@ class AbstractDatasetResource(ABC):
                     f"A more mature version of dataset {ds.id} already exists, with id: "
                     f"{dupe.id} and maturity: {dupe.metadata.dataset_maturity}"
                 )
-            less_mature.append(dupe.id)
-        self.archive(less_mature)
-        for lm_ds in less_mature:
-            _LOG.info(f"Archived less mature dataset: {lm_ds}")
+            less_mature.append(dupe)
+        return less_mature
 
     @abstractmethod
     def restore(self, ids: Iterable[DSID]) -> None: