+++ integration_tests/index/test_index_data.py
@@ -94,7 +94,6 @@ def test_archive_datasets(index, local_config, ls8_eo3_dataset):
     assert not indexed_dataset.is_archived
 
 
-@pytest.mark.parametrize('datacube_env_name', ('experimental',))
 def test_archive_less_mature(index, final_dataset, nrt_dataset):
     # case 1: add nrt then final; nrt should get archived
     index.datasets.add(nrt_dataset, with_lineage=False, archive_less_mature=True)
@@ -109,6 +108,15 @@ def test_archive_less_mature(index, final_dataset, nrt_dataset):
     with pytest.raises(ValueError):
         # should error as more mature version of dataset already exists
         index.datasets.add(nrt_dataset, with_lineage=False, archive_less_mature=True)
+
+
+def test_archive_less_mature_approx_timestamp(index, ga_s2am_ard3_final, ga_s2am_ard3_interim):
+    # test archive_less_mature where there's a slight difference in timestamps
+    index.datasets.add(ga_s2am_ard3_interim, with_lineage=False)
+    index.datasets.get(ga_s2am_ard3_interim.id).is_active
+    index.datasets.add(ga_s2am_ard3_final, with_lineage=False, archive_less_mature=True)
+    assert index.datasets.get(ga_s2am_ard3_interim.id).is_archived
+    assert index.datasets.get(ga_s2am_ard3_final.id).is_active
 
 
 def test_purge_datasets(index, ls8_eo3_dataset):